% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{adjust_overlap}
\alias{adjust_overlap}
\title{Automatically adjust overlap parameter for ICP}
\usage{
adjust_overlap(overlap, r, M)
}
\arguments{
\item{overlap}{default value}

\item{r}{plot radius}

\item{M}{brute force registration matrix}
}
\description{
Automatically adjust overlap parameter for ICP
}
