% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{compare_alignment}
\alias{compare_alignment}
\title{Compare the Alignment of Two Point Clouds}
\usage{
compare_alignment(pc1, pc2)
}
\arguments{
\item{pc1}{A matrix of XYZ coordinates or a `LAS` object from the `lidR` package.}

\item{pc2}{A matrix of XYZ coordinates or a `LAS` object from the `lidR` package.}
}
\description{
Computes the root mean square error (RMSE) of the nearest neighbor distances between two point clouds.
This function measures how well `pc2` aligns with `pc1` by finding the closest points in `pc1` for
each point in `pc2` and computing the RMSE for half of the point cloud. It uses the 50% of the
most aligned points to account for non-overlapping features. Thus, the returned value is not
a true RMSE but rather an arbitrary alignment score.
}
