% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features.R
\name{extract_features}
\alias{extract_features}
\title{Extract Features from a LiDAR Point Cloud}
\usage{
extract_features(las, strategy = "chm-dtm")
}
\arguments{
\item{las}{A \code{LAS} object containing the LiDAR point cloud data.}

\item{strategy}{Character. The extraction strategy to use. \code{"chm-dtm"} or \code{"mls-tls"}. See details.}
}
\value{
A \code{LAS} object containing the extracted features.
}
\description{
This function extracts features from a LiDAR point cloud based on the specified strategy in order to
retain points that are likely matching in order to perform a registration of the point clouds.
}
\details{
\itemize{
\item The \code{"chm-dtm"} strategy extracts a digital terrain model (DTM) and a canopy height model (CHM),
and merges them into a new \code{LAS} object.
\item The \code{"mls-tls"} strategy applies a series of filtering and processing steps including noise removal,
smoothing, decimation, anisotropy filtering, and connected component filtering in order to extract
trees in the first 4 meters of the point clouds
}
}
