% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{find_cloudcompare}
\alias{find_cloudcompare}
\title{Find CloudCompare executable path}
\usage{
find_cloudcompare()
}
\value{
A character string with the path to the CloudCompare executable.
}
\description{
This function attempts to locate the CloudCompare executable based on the operating system.
On Windows, it queries the registry to find the installation path. On macOS and Linux (Flathub version),
it returns an error.
}
