% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{icp}
\alias{icp}
\title{CloudCompare ICP Registration}
\usage{
icp(
  vref,
  umov,
  min_error_diff = 1e-05,
  overlap = 90,
  rot = "Z",
  skip_txy = FALSE,
  cc = find_cloudcompare(),
  verbose = TRUE
)
}
\arguments{
\item{vref}{A \code{LAS} containing the reference point cloud}

\item{umov}{A \code{LAS} containing the moving point cloud}

\item{min_error_diff}{Numeric. The minimum error difference for the ICP algorithm to terminate (default: 1e-5).}

\item{overlap}{Numeric. to specify the percentage of (final) overlap (integer number between 10 and 100 - default = 100)}

\item{rot}{Character. 'XYZ' or 'X' or 'Y' or 'Z' or 'NONE' to constrain the rotation around a given axis}

\item{skip_txy}{Logical. Block XY translation.
(or no rotation at all)}

\item{cc}{Character. The path to the CloudCompare executable}

\item{verbose}{logical}
}
\value{
A 4x4 transformation matrix resulting from the ICP registration.
}
\description{
This function performs point cloud registration using the CloudCompare ICP (Iterative Closest Point) algorithm.
CloudCompare must be installed to use this function. You can download it from: https://www.danielgm.net/cc/.
}
\details{
The function exports both point clouds to temporary \code{.xyz} files, runs CloudCompare ICP
via a system call, retrieves the transformation matrix from the generated file, and then deletes
temporary files.
}
