% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Create a Rotation Matrix}
\usage{
rotation_matrix(angle, axis = "z")
}
\arguments{
\item{angle}{Rotation angle in degrees.}

\item{axis}{Axis of rotation ('x', 'y', or 'z').}
}
\value{
A 4x4 transformation matrix.
}
\description{
Generates a 4x4 homogeneous transformation matrix for rotation around X, Y, or Z.
}
