% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{rtm_info}
\alias{rtm_info}
\title{Extract Translation and Rotation from a Transformation Matrix}
\usage{
rtm_info(M)
}
\arguments{
\item{M}{A 4x4 transformation matrix.}
}
\value{
None (prints results).
}
\description{
Prints translation and rotation angles from a 4x4 transformation matrix.
}
