% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{show_alignment}
\alias{show_alignment}
\title{Visualize Alignment of Point Clouds}
\usage{
show_alignment(ref, mov, M = NULL, size = 2)
}
\arguments{
\item{ref, mov}{LAS objects}

\item{M}{A 4x4 transformation matrix (optional). If provided, the moving dataset is transformed accordingly.}

\item{size}{Numeric, point size for visualization. Default is 2.}
}
\value{
A 3D plot of the reference and transformed moving point clouds. If a transformation is applied,
translation and rotation are also visualized with lines.
}
\description{
This function visualizes the alignment of a moving (\code{mov}) point cloud with a reference (\code{ref}) point cloud.
If a transformation matrix (\code{M}) is provided, the function applies the transformation and displays
the translation and rotation.
}
