% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{translate_las}
\alias{translate_las}
\title{Apply Translation to a LAS Object}
\usage{
translate_las(las, tx = 0, ty = 0, tz = 0)
}
\arguments{
\item{las}{A \code{LAS} object.}

\item{tx}{Translation along X-axis (default 0).}

\item{ty}{Translation along Y-axis (default 0).}

\item{tz}{Translation along Z-axis (default 0).}
}
\value{
A modified LAS object.
}
\description{
Translates a LAS point cloud by a given offset.
}
