% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{translation_matrix}
\alias{translation_matrix}
\title{Create a Translation Matrix}
\usage{
translation_matrix(tx, ty, tz)
}
\arguments{
\item{tx}{Translation along the X-axis.}

\item{ty}{Translation along the Y-axis.}

\item{tz}{Translation along the Z-axis.}
}
\value{
A 4x4 transformation matrix.
}
\description{
Generates a 4x4 homogeneous transformation matrix for translation.
}
