% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{AlignmentScene}
\alias{AlignmentScene}
\title{AlignmentScene R6 Class for Point Cloud Alignment}
\description{
The `AlignmentScene` R6 class provides methods for aligning two 3D point clouds plots, typically
from terrestrial (MLS/TLS) or aerial (ALS/UAV) LiDAR sources. It implements a multi-step alignment
process with coarse, fine, and extra-fine alignment stages.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AlignmentScene-new}{\code{AlignmentScene$new()}}
\item \href{#method-AlignmentScene-print}{\code{AlignmentScene$print()}}
\item \href{#method-AlignmentScene-set_radius}{\code{AlignmentScene$set_radius()}}
\item \href{#method-AlignmentScene-set_ref_is_ground_based}{\code{AlignmentScene$set_ref_is_ground_based()}}
\item \href{#method-AlignmentScene-set_mov_is_ground_based}{\code{AlignmentScene$set_mov_is_ground_based()}}
\item \href{#method-AlignmentScene-prepare}{\code{AlignmentScene$prepare()}}
\item \href{#method-AlignmentScene-coarse_align}{\code{AlignmentScene$coarse_align()}}
\item \href{#method-AlignmentScene-fine_align}{\code{AlignmentScene$fine_align()}}
\item \href{#method-AlignmentScene-extra_fine_align}{\code{AlignmentScene$extra_fine_align()}}
\item \href{#method-AlignmentScene-align}{\code{AlignmentScene$align()}}
\item \href{#method-AlignmentScene-plot}{\code{AlignmentScene$plot()}}
\item \href{#method-AlignmentScene-get_registration_matrix}{\code{AlignmentScene$get_registration_matrix()}}
\item \href{#method-AlignmentScene-clone}{\code{AlignmentScene$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-new"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-new}{}}}
\subsection{Method \code{new()}}{
Create a new AlignmentScene object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$new(fref, fmov, radius = 20, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fref, fmov}}{character, path to the las laz file with the point cloud to align. fref is
the reference point cloud, fmov is the point cloud to be aligned}

\item{\code{radius}}{clip radius. The routine does not align the full point cloud but clip before
to compute the registration matrix}

\item{\code{verbose}}{logical}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `AlignmentScene` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-print"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-print}{}}}
\subsection{Method \code{print()}}{
print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-set_radius"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-set_radius}{}}}
\subsection{Method \code{set_radius()}}{
Set clip radius
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$set_radius(radius)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{radius}}{numeric.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-set_ref_is_ground_based"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-set_ref_is_ground_based}{}}}
\subsection{Method \code{set_ref_is_ground_based()}}{
Tell the pipeline if the reference point cloud is ground-based (TLS, MLS)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$set_ref_is_ground_based(val = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{logical}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-set_mov_is_ground_based"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-set_mov_is_ground_based}{}}}
\subsection{Method \code{set_mov_is_ground_based()}}{
Tell the pipeline if the moving point cloud is ground-based (TLS, MLS)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$set_mov_is_ground_based(val = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{logical}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-prepare"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-prepare}{}}}
\subsection{Method \code{prepare()}}{
First function to run. It read the point cloud, extract features, and perform raw alignment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$prepare()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-coarse_align"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-coarse_align}{}}}
\subsection{Method \code{coarse_align()}}{
Second function to run. It perform a brute force alignment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$coarse_align(res = 2, max_offset = 10, debug = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{res}}{numeric. Subsampling resolution. Keep it as is.}

\item{\code{max_offset}}{numeric. Maximum translation possible. Increase if the point cloud are
strongly missaligned on XY.}

\item{\code{debug}}{logical.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-fine_align"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-fine_align}{}}}
\subsection{Method \code{fine_align()}}{
Third function to run. It perform an ICP alignment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$fine_align(overlap = "auto", use_cc = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{overlap}}{numeric trimmed ICP overlap. Can be 10, 20, 30, 40 to 100.}

\item{\code{use_cc}}{bool. Use CloudCompare ICP instead of native ICP.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-extra_fine_align"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-extra_fine_align}{}}}
\subsection{Method \code{extra_fine_align()}}{
Fourth function to run. It perform an ICP alignment on the trunks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$extra_fine_align()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-align"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-align}{}}}
\subsection{Method \code{align()}}{
Full pipeline
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$align(res = 2, max_offset = 8)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{res}}{numeric. Subsampling resolution. Keep it as is.}

\item{\code{max_offset}}{numeric. Maximum translation possible. Increase if the point cloud are
strongly missaligned on XY.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-plot"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-plot}{}}}
\subsection{Method \code{plot()}}{
Plot alignment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$plot(
  which = c("raw", "coarse", "fine", "extra"),
  compare_to = which
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which}}{string. The alignment level to plot}

\item{\code{compare_to}}{string. The alignment level to plot and to compare}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-get_registration_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-get_registration_matrix}{}}}
\subsection{Method \code{get_registration_matrix()}}{
Get the registration matrix. This matrix is more or less accurate as a function of the level
of registration performed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$get_registration_matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlignmentScene-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AlignmentScene-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlignmentScene$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
