% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudcompare.R
\name{find_cloudcompare}
\alias{find_cloudcompare}
\alias{register_cloudcompare}
\title{Find CloudCompare executable path}
\usage{
find_cloudcompare()

register_cloudcompare(path)
}
\arguments{
\item{path}{A string with full path to the CloudCompare binary}
}
\value{
A character string with the path to the CloudCompare executable.
}
\description{
This function attempts to locate the CloudCompare executable based on the operating system.
On Windows, it queries the registry to find the installation path. On macOS and Linux (Flathub version),
it returns an error. If the path is not found automatically use `register_cloudcompare()`
}
